Moved simple functions to tools file
